feat(FE): 사용자 프로필 편집을 위한 닉네임 및 프로필 이미지 폼 추가
- 사용자 프로필 편집을 간소화하기 위해 `NicknameForm`과 `ProfileImageForm` 컴포넌트를 도입
- 코드 조직과 가독성을 향상시키기 위해 `EditUserProfile` 컴포넌트를 리팩토링하여 새로운 폼을 활용
- 회원가입 시 사용자 경험을 개선하기 위해 온보딩 과정에 프로필 이미지와 닉네임 입력 컴포넌트 사용
